+++ changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - The scout:index and scout:delete commands are now implemented for Elasticsearch
 - A DocumentAdapterInterface with an adapter that only deals with documents
+- The Query String syntax
 
 ### Changed
 - The IndexAdapterInterface now only focuses on indices